+++ mozetl/taar/taar_ensemble.py
@@ -1,196 +1,35 @@
-import boto3
-import json
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this file,
+# You can obtain one at http://mozilla.org/MPL/2.0/.
+
+import boto3  # noqa
+import json        # noqa
+import itertools        # noqa
+import copy
+import numpy        # noqa
+from botocore.exceptions import ClientError        # noqa
+from random import seed        # noqa
+from random import randrange        # noqa
+from .xvalidator import XValidator        # noqa
+
+from taar.recommenders import CollaborativeRecommender
+from taar.recommenders import LegacyRecommender
+from taar.recommenders import LocaleRecommender
+from taar.recommenders import RecommendationManager
+from taar.recommenders import SimilarityRecommender
+
 import logging
-import numpy
-from botocore.exceptions import ClientError
-from csv import reader
-from math import sqrt
-from math import exp
-from pyspark.sql import SparkSession
-from pyspark.sql.functions import col
-from pyspark.sql.functions import *
-from random import seed
-from random import randrange
 
-
-# Helper function to do some math for cllr.
-def neg_log_sig(log_odds):
-    neg_log_odds = [-1.0 * x for x in log_odds]
-    e = numpy.exp(neg_log_odds)
-    return [numpy.log(1 + f) for f in e if f < (f + 1)]
-
-
-# Compute the log likelihood ratio cost which should be minimized.
-def cllr(lrs_on_target, lrs_off_target):
-    # based on Niko Brummer's original implementation:
-    # Niko Brummer and Johan du Preez, Application-Independent Evaluation of Speaker Detection"
-    # Computer Speech and Language, 2005
-    lrs_on_target = numpy.log(lrs_on_target[~numpy.isnan(lrs_on_target)])
-    lrs_off_target = numpy.log(lrs_off_target[~numpy.isnan(lrs_off_target)])
-
-    c1 = numpy.mean(neg_log_sig(lrs_on_target)) / numpy.log(2)
-    c2 = numpy.mean(neg_log_sig(-1.0 * lrs_off_target)) / numpy.log(2)
-    return (c1 + c2) / 2
-
-
-def evalcllr(recommendations_list, unmasked_addons):
-    """ A helper function to evaluate the performance of a particular recommendation 
-    strategy on a client with a set of installed addons that have been patially masked.
-    Keyword arguments:
-    recommendations_list -- a list of tuples containing (guid, confidence) pairs.
-    unmasked_addons -- a list of the true installed addon guids for a test client.
-    """
-    # Organizer function to extract weights from recommendation list for passing to cllr.
-    lrs_on_target_helper = [item[1] for item in recommendations_list if item[0] in unmasked_addons]
-    lrs_off_target_helper = [item[1] for item in recommendations_list if item[0] not in unmasked_addons]
-    return cllr(lrs_on_target_helper, lrs_off_target_helper)
-
-
-def load_training_from_telemetry():
-    # Define the set of feature names to be used in the donor computations.
-    AMO_DUMP_BUCKET = 'telemetry-parquet'
-    AMO_DUMP_KEY = 'telemetry-ml/addon_recommender/addons_database.json'
-
-    def load_amo_external_whitelist():
-        """ Download and parse the AMO add-on whitelist.
-        The json fetched here is generated by a weekly job that queries the AMO public API
-        this whitelist is sure to exclude any Mozilla studies that are not correclty filtered
-        as well as personal/unsigned/unlisted addons manually installed.
-        :raises RuntimeError: the AMO whitelist file cannot be downloaded or contains
-                              no valid add-ons.
-        """
-        final_whitelist = []
-        amo_dump = {}
-        try:
-            # Load the most current AMO dump JSON resource.
-            s3 = boto3.client('s3')
-            s3_contents = s3.get_object(Bucket=AMO_DUMP_BUCKET, Key=AMO_DUMP_KEY)
-            amo_dump = json.loads(s3_contents['Body'].read())
-        except ClientError:
-            logger.exception("Failed to download from S3", extra={
-                "bucket": AMO_DUMP_BUCKET,
-                "key": AMO_DUMP_KEY})
-
-        # If the load fails, we will have an empty whitelist, this may be problematic.
-        for key, value in amo_dump.items():
-            addon_files = value.get('current_version', {}).get('files', {})
-            # If any of the addon files are web_extensions compatible, it can be recommended.
-            if any([f.get("is_webextension", False) for f in addon_files]):
-                final_whitelist.append(value['guid'])
-
-        if len(final_whitelist) == 0:
-            raise RuntimeError("Empty AMO whitelist detected")
-
-        return set(final_whitelist)
-
-    def get_initial_sample():
-        """ Takes an initial sample from the longitudinal dataset 
-        (randomly sampled from main summary). Coarse filtering on:
-        - number of installed addons
-        - corrupt and generally wierd telemetry entries
-        - isolating release channel
-        - column selection
-        """
-        client_features_frame = spark.sql("SELECT * FROM longitudinal")\
-                .where("active_addons IS NOT null")\
-                .where("size(active_addons[0]) > 2")\
-                .where("size(active_addons[0]) < 100")\
-                .where("normalized_channel = 'release'")\
-                .where("build IS NOT NULL AND build[0].application_name = 'Firefox'")\
-                .selectExpr(\
-                    "client_id as client_id",\
-                    "active_addons[0] as active_addons",\
-                    "geo_city[0] as geo_city",\
-                    "subsession_length[0] as subsession_length",\
-                    "settings[0].locale as locale",\
-                    "os as os",\
-                    "places_bookmarks_count[0].sum AS bookmark_count",\
-                    "scalar_parent_browser_engagement_tab_open_event_count[0].value AS tab_open_count",\
-                    "scalar_parent_browser_engagement_total_uri_count[0].value AS total_uri",\
-                    "scalar_parent_browser_engagement_unique_domains_count[0].value AS unique_tlds")
-        return client_features_frame
-
-    def get_addons_per_client(users_df, minimum_addons_count):
-        """ Extracts a DataFrame that contains one row
-        for each client along with the list of active add-on GUIDs.
-        """
-        def is_valid_addon(guid, addon):
-            """ Filter individual addons out to exclude, system addons,
-            legacy addons, disabled addons, sideloaded addons.
-            """
-            return not (
-                addon.is_system or
-                addon.app_disabled or
-                addon.type != "extension" or
-                addon.user_disabled or
-                addon.foreign_install or
-                guid not in broadcast_amo_whitelist.value
-            )
-        # may need addiitonal whitelisting to remove shield addons
-
-        # Create an add-ons dataset un-nesting the add-on map from each
-        # user to a list of add-on GUIDs. Also filter undesired add-ons.
-        return (
-            users_df.rdd
-            .map(lambda p: (p["client_id"],
-                 [guid for guid, data in p["active_addons"].items() if is_valid_addon(guid, data)]))
-            .filter(lambda p: len(p[1]) > minimum_addons_count)
-            .toDF(["client_id", "addon_ids"])
-        )
-
-    # TODO: Add logging at each of these operations. 
-    client_features_frame = get_initial_sample()
-
-    amo_white_list = load_amo_external_whitelist()
-    broadcast_amo_whitelist = sc.broadcast(amo_white_list)
-
-    addons_info_frame = get_addons_per_client(client_features_frame, 4)
-    taar_training = addons_info_frame.join(client_features_frame, 'client_id', 'inner').drop('active_addons')
-
-    return taar_training
-
-# All this is not needed, delete after getting a functional version
-# Convert string column to float
-# def str_column_to_float(dataset, column):
-#    for row in dataset:
-#        row[column] = float(row[column].strip())
-
-
-# Convert string column to integer
-# def str_column_to_int(dataset, column):
-#    class_values = [row[column] for row in dataset]
-#    unique = set(class_values)
-#    lookup = dict()
-#    for i, value in enumerate(unique):
-#        lookup[value] = i
-#    for row in dataset:
-#        row[column] = lookup[row[column]]
-#    return lookup
-
-
-# Split a dataset into k folds
-def cross_validation_split(dataset, n_folds):
-    # TODO: inside the cross validation we also need to handle the masking of installed addons for the training set.
-    # one way to do this would be to precompute a binary mask for each chunk, or implement this in the
-    # "evaluate_algorithm" function.
-    dataset_split = list()
-    dataset_copy = list(dataset)
-    fold_size = int(len(dataset) / n_folds)
-    for fold_i in range(n_folds):
-        fold = list()
-        while len(fold) < fold_size:
-            index = randrange(len(dataset_copy))
-            fold.append(dataset_copy.pop(index))
-        dataset_split.append(fold)
-    return dataset_split
+logger = logging.getLogger(__name__)
 
 
 # Calculate accuracy percentage - substitute accuracy percentage for cllr
 # We could use 1.0-cllr as a metric to reuse even more existing code.
 # This will be the cllr code now.
-# TODO: include the (prediction, confidence) tuple in the INPUT vector so that evalCLLR can be called
-# TODO: still include the input vector "actual" which will be passed to the evalcllr function as
-# unmasked_addons.
+# TODO: include the (prediction, confidence) tuple in the INPUT vector
+# so that eval_cllr can be called
+# TODO: still include the input vector "actual" which will be passed
+# to the eval_cllr function as unmasked_addons.
 def accuracy_metric(actual, predicted):
     correct = 0
     for i in range(len(actual)):
@@ -201,125 +40,61 @@ def accuracy_metric(actual, predicted):
 
 # Evaluate an algorithm using a cross validation split
 def evaluate_algorithm(dataset, algorithm, n_folds, *args):
-    # TODO: here the algorithm argument shoudl be a list of taar modules.
-    folds = cross_validation_split(dataset, n_folds)
+    # TODO: here the algorithm argument should be a list of taar modules.
+
+    # TODO: copy MIN_ADDONSETSIZE from *args
+    MIN_ADDONSETSIZE = 2
+    xvalidator = XValidator(n_folds, MIN_ADDONSETSIZE)
+    folds = xvalidator.cross_validation_split(dataset)
+
     scores = list()
     for fold in folds:
+
+        # Create a training set which is the entire list of folds,
+        # less the current fold
         train_set = list(folds)
         train_set.remove(fold)
-        train_set = sum(train_set, [])
-        test_set = list()
-        for row in fold:
-            row_copy = list(row)
-            test_set.append(row_copy)
-            row_copy[-1] = None
-        predicted = algorithm(train_set, test_set, *args)
+
+        # train_set needs to be converted from a list of lists into a
+        # single monolithic list
+        train_set = itertools.chain.from_iterable(train_set)
+        test_set = copy.copy(fold)
+
         # TODO this should just call the appropriate TAAR module (collaborative, similarity, or locale)
         # and get the recommendations over the (masked) training clients.
         # 1- The unmasked installed addons in that client is "actual"
         # 2- the output of the TAAR module that was called is "predicted"
+        predicted = algorithm(train_set, test_set, *args)
+
         actual = [row[-1] for row in fold]
         accuracy = accuracy_metric(actual, predicted)
         # TODO: accuracy will now be CLLR
         scores.append(accuracy)
     return scores
 
-# ALL THIS SHIT NEEDS TO GO:
-
-# # Calculate the Euclidean distance between two vectors
-# def euclidean_distance(row1, row2):
-#     distance = 0.0
-#     for i in range(len(row1) - 1):
-#         distance += (row1[i] - row2[i]) ** 2
-#     return sqrt(distance)
-#
-#
-# # Locate neighbors for a new row
-# def get_neighbors(train, test_row, num_neighbors):
-#     distances = list()
-#     for train_row in train:
-#         dist = euclidean_distance(test_row, train_row)
-#         distances.append((train_row, dist))
-#     distances.sort(key=lambda tup: tup[1])
-#     neighbors = list()
-#     for i in range(num_neighbors):
-#         neighbors.append(distances[i][0])
-#     return neighbors
-#
-#
-# # Make a prediction with kNN
-# def knn_predict(model, test_row, num_neighbors=2):
-#     neighbors = get_neighbors(model, test_row, num_neighbors)
-#     output_values = [row[-1] for row in neighbors]
-#     prediction = max(set(output_values), key=output_values.count)
-#     return prediction
-#
-#
-# # Prepare the kNN model
-# def knn_model(train):
-#     return train
-#
-#
-# # Make a prediction with weights
-# def perceptron_predict(model, row):
-#     activation = model[0]
-#     for i in range(len(row) - 1):
-#         activation += model[i + 1] * row[i]
-#     return 1.0 if activation >= 0.0 else 0.0
-#
-#
-# # Estimate Perceptron weights using stochastic gradient descent
-# def perceptron_model(train, l_rate=0.01, n_epoch=5000):
-#     weights = [0.0 for i in range(len(train[0]))]
-#     for epoch in range(n_epoch):
-#         for row in train:
-#             prediction = perceptron_predict(weights, row)
-#             error = row[-1] - prediction
-#             weights[0] = weights[0] + l_rate * error
-#             for i in range(len(row) - 1):
-#                 weights[i + 1] = weights[i + 1] + l_rate * error * row[i]
-#     return weights
-#
-#
-# # Make a prediction with coefficients
-# def logistic_regression_predict(model, row):
-#     yhat = model[0]
-#     for i in range(len(row) - 1):
-#         yhat += model[i + 1] * row[i]
-#     return 1.0 / (1.0 + exp(-yhat))
-#
-#
-# # Estimate logistic regression coefficients using stochastic gradient descent
-# def logistic_regression_model(train, l_rate=0.01, n_epoch=5000):
-#     coef = [0.0 for i in range(len(train[0]))]
-#     for epoch in range(n_epoch):
-#         for row in train:
-#             yhat = logistic_regression_predict(coef, row)
-#             error = row[-1] - yhat
-#             coef[0] = coef[0] + l_rate * error * yhat * (1.0 - yhat)
-#             for i in range(len(row) - 1):
-#                 coef[i + 1] = coef[i + 1] + l_rate * error * yhat * (1.0 - yhat) * row[i]
-#     return coef
-
 
 # Make predictions with sub-models and construct a new stacked row
 def to_stacked_row(models, predict_list, row):
     stacked_row = list()
-    for i in range(len(models)):
-        prediction = predict_list[i](models[i], row)
+    for model in models:
+        prediction = predict_list[i](model, row)
         stacked_row.append(prediction)
     stacked_row.append(row[-1])
-    return row[0:len(row) - 1] + stacked_row
+    return row[:len(row) - 1] + stacked_row
 
 
 # Stacked Generalization Algorithm
-# TODO: this is where the actual work is, we need to get a comparable flow 
-# utilizing the TAAR models as with these on the standard ML models, 
-# the gradient decent algorithm should iterate to a set of meta-parameters 
+# TODO: this is where the actual work is, we need to get a comparable flow
+# utilizing the TAAR models as with these on the standard ML models,
+# the gradient decent algorithm should iterate to a set of meta-parameters
 # weighting the recommneder modules.
 def stacking(train, test):
-    # TODO model list now refers to instantiated TAAR recommender modules.
-    model_list = [knn_model, perceptron_model]
+    # model list now refers to instantiated TAAR recommender modules.
+    model_list = [CollaborativeRecommender(), LegacyRecommender(),
+                  LocaleRecommender(), SimilarityRecommender(),
+                  RecommendationManager()]
+
+
     predict_list = [knn_predict, perceptron_predict]
     models = list()
     for i in range(len(model_list)):
@@ -339,6 +114,19 @@ def stacking(train, test):
         predictions.append(prediction)
     return predictions
 
+
+# Estimate logistic regression coefficients using stochastic gradient descent
+def logistic_regression_model(train, l_rate=0.01, n_epoch=5000):
+    coef = [0.0 for i in range(len(train[0]))]
+    for epoch in range(n_epoch):
+        for row in train:
+            yhat = logistic_regression_predict(coef, row)
+            error = row[-1] - yhat
+            coef[0] = coef[0] + l_rate * error * yhat * (1.0 - yhat)
+            for i in range(len(row) - 1):
+                coef[i + 1] = coef[i + 1] + l_rate * error * yhat * (1.0 - yhat) * row[i]
+    return coef
+-
 
 # Test stacking on the sonar dataset
 seed(1)
